code refactoring

+++ src/test/java/base/BaseTest.java
@@ -7,6 +7,9 @@ import org.testng.annotations.AfterSuite;
 import org.testng.annotations.AfterTest;
 import pages.*;
 
+import static util.GetBoolean.getBooleanFromString;
+import static util.PropertiesUtil.getProp;
+
 public class BaseTest {
     protected WebDriver driver = CommonActions.createDriver();
     protected BasePage basePage = new BasePage(driver);
@@ -17,13 +20,19 @@ public class BaseTest {
 
     @AfterTest
     public void clearCookiesAndLocalStorage() {
-        JavascriptExecutor jsExecutor = (JavascriptExecutor) driver;
-        driver.manage().deleteAllCookies();
-        jsExecutor.executeScript("window.sessionStorage.clear();");
+        boolean clearCookiesAndStorage = getBooleanFromString(getProp("clear_cookies_and_storage"));
+        if (clearCookiesAndStorage) {
+            JavascriptExecutor jsExecutor = (JavascriptExecutor) driver;
+            driver.manage().deleteAllCookies();
+            jsExecutor.executeScript("window.sessionStorage.clear();");
+        }
     }
 
     @AfterSuite(alwaysRun = true)
     public void close() {
-        driver.quit();
+        boolean closeBrowser = getBooleanFromString(getProp("close_browser"));
+        if (closeBrowser) {
+            driver.quit();
+        }
     }
 }
